Fix three bugs with the s3 sync.
One was that a string instead of a buffer was passed to knox, resulted in an incorrect Content-Length header and thus truncated html files.
The other being that the hashing for the html files never worked, since we compared the hash of the string against the hash of the file.
The third, that the files that are included from multiple pages were always uploaded at least once, since the hash was deleted from the list after the first existing
version was detected.